Delete some unnessesary classes and add models namespace
