+++ src/dataModel/CurrencyConverter.java
@@ -1,5 +1,7 @@
 package dataModel;
 
+import dataModel.models.Currency;
+
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -10,7 +12,7 @@ import java.util.TreeMap;
  * A class that supports currency conversion. A "currency value" is a relative
  * value of a currency. The exchange rate between a pair of currency types is
  * based on the relative value of each currency. The class uses a
- * {@link TreeMap<Currency, BigDecimal>} as internal representation. All class
+ * {@link TreeMap< Currency , BigDecimal>} as internal representation. All class
  * methods are synchronized (towards {@link CurrencyConverter#currencyValues} as
  * necessary to support concurrent calls. All {@link CurrencyConverter} objects:
  * <p>
